backend/requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiomqtt==2.3.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fbchat-muqit==1.1.2
frozenlist==1.5.0
GeoAlchemy2==0.17.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.2.0
openai==1.68.2
packaging==24.2
paho-mqtt==2.1.0
propcache==0.3.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
